ワーシャルフロイドを分離
